backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.4.0
django-filter>=24.2
python-decouple>=3.8
reportlab>=4.2.0
psycopg2-binary>=2.9.9
requests>=2.32.0
cryptography>=43.0.0
